# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AiSOC Platform API
  description: |
    Open-source AI Security Operations Center (v5.2.0).

    Autonomous threat detection, investigation, and response — with UEBA behavioral
    analytics, honeytoken deception, purple-team adversary emulation, enterprise RBAC,
    immutable audit logging, the v5.2 Investigation Ledger (signed agent actions),
    Ambient Copilot, Responder PWA, public evaluation harness, MCP server, and
    SOC 2 / ISO 27001 / NIST CSF compliance dashboards.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained from `POST /api/v1/auth/login`.
    Service-to-service calls may use an API key via `X-API-Key` header.

    ## Multi-tenancy
    All data is isolated via PostgreSQL Row-Level Security. The `tenant_id` is derived
    from the authenticated token and cannot be overridden by the caller.

    ## Rate limiting
    Default: 1000 req/min per API key; 100 req/min per user token.
    Limits are surfaced via `X-RateLimit-*` response headers.
  version: 5.2.0
  contact:
    name: AiSOC Community
    url: https://github.com/beenuar/AiSOC/security/advisories/new
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
paths:
...
  /api/v1/graph/attack-path/{case_id}:
    get:
      tags:
      - graph
      summary: Get attack path graph for a case
      description: 'Traverse the knowledge graph from a Case node to reconstruct the
        full

        attack path: Case → Alerts → Hosts/Users → IOCs → MITRE Techniques.'
      operationId: get_attack_path_api_v1_graph_attack_path__case_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: max_depth
        in: query
        required: false
        schema:
          type: integer
          maximum: 10
          minimum: 1
          default: 6
          title: Max Depth
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttackPathResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graph/blast-radius/{entity_type}/{entity_id}:
    get:
      tags:
      - graph
      summary: Compute blast radius from an entity
      description: 'Compute the blast radius starting from a Host, User, or IOC node.

        Returns all entities reachable within `hops` and a severity score.'
      operationId: get_blast_radius_api_v1_graph_blast_radius__entity_type___entity_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: entity_type
        in: path
        required: true
        schema:
          type: string
          title: Entity Type
      - name: entity_id
        in: path
        required: true
        schema:
          type: string
          title: Entity Id
      - name: hops
        in: query
        required: false
        schema:
          type: integer
          maximum: 6
          minimum: 1
          default: 3
          title: Hops
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlastRadiusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graph/neighbors/{entity_type}/{entity_id}:
    get:
      tags:
      - graph
      summary: Get immediate graph neighbors of an entity
      description: Return all nodes directly connected (depth 1) to the specified
        entity.
      operationId: get_entity_neighbors_api_v1_graph_neighbors__entity_type___entity_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: entity_type
        in: path
        required: true
        schema:
          type: string
          title: Entity Type
      - name: entity_id
        in: path
        required: true
        schema:
          type: string
          title: Entity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityNeighborsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/graph/mitre-coverage:
    get:
      tags:
      - graph
      summary: MITRE ATT&CK technique coverage for tenant
      description: Return MITRE ATT&CK technique coverage aggregated from all tenant
        alerts.
      operationId: get_mitre_coverage_api_v1_graph_mitre_coverage_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MitreCoverageItem'
                type: array
                title: Response Get Mitre Coverage Api V1 Graph Mitre Coverage Get
      security:
      - HTTPBearer: []
  /api/v1/graph/entities/host:
    post:
      tags:
      - graph
      summary: Upsert a Host node in the graph
      description: Create or update a Host node in the knowledge graph.
      operationId: upsert_host_api_v1_graph_entities_host_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertHostRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Upsert Host Api V1 Graph Entities Host Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/graph/entities/user:
    post:
      tags:
      - graph
      summary: Upsert a User node in the graph
      description: Create or update a User node in the knowledge graph.
      operationId: upsert_user_api_v1_graph_entities_user_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertUserRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Upsert User Api V1 Graph Entities User Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/graph/entities/alert:
    post:
      tags:
      - graph
      summary: Upsert an Alert node and its relationships in the graph
      description: 'Create or update an Alert node and link it to Host, User, IOC,
        and Technique nodes.

        Called automatically after alert creation to keep the graph in sync.'
      operationId: upsert_alert_graph_api_v1_graph_entities_alert_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertAlertGraphRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Upsert Alert Graph Api V1 Graph Entities Alert Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/graph/entities/case:
    post:
      tags:
      - graph
      summary: Upsert a Case node and link to alerts in the graph
      description: Create or update a Case node and link it to Alert nodes.
      operationId: upsert_case_graph_api_v1_graph_entities_case_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCaseGraphRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Upsert Case Graph Api V1 Graph Entities Case Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    AttackPathResponse:
      properties:
        case_id:
          type: string
          title: Case Id
        nodes:
          items:
            $ref: '#/components/schemas/GraphNode'
          type: array
          title: Nodes
        edges:
          items:
            $ref: '#/components/schemas/GraphEdge'
          type: array
          title: Edges
        node_count:
          type: integer
          title: Node Count
        edge_count:
          type: integer
          title: Edge Count
      type: object
      required:
      - case_id
      - nodes
      - edges
      - node_count
      - edge_count
      title: AttackPathResponse
    BlastRadiusResponse:
      properties:
        entity_id:
          type: string
          title: Entity Id
        entity_type:
          type: string
          title: Entity Type
        hops:
          type: integer
          title: Hops
        affected_nodes:
          items:
            $ref: '#/components/schemas/GraphNode'
          type: array
          title: Affected Nodes
        total_affected:
          type: integer
          title: Total Affected
        type_breakdown:
          additionalProperties:
            type: integer
          type: object
          title: Type Breakdown
        blast_radius_score:
          type: number
          title: Blast Radius Score
      type: object
      required:
      - entity_id
      - entity_type
      - hops
      - affected_nodes
      - total_affected
      - type_breakdown
      - blast_radius_score
      title: BlastRadiusResponse
...
    EntityNeighborsResponse:
      properties:
        entity_id:
          type: string
          title: Entity Id
        entity_type:
          type: string
          title: Entity Type
        source:
          anyOf:
          - $ref: '#/components/schemas/GraphNode'
          - type: 'null'
        neighbors:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Neighbors
        neighbor_count:
          type: integer
          title: Neighbor Count
          default: 0
      type: object
      required:
      - entity_id
      - entity_type
      - source
      - neighbors
      title: EntityNeighborsResponse
...
    GraphEdge:
      properties:
        source:
          type: string
          title: Source
        target:
          type: string
          title: Target
        type:
          type: string
          title: Type
      type: object
      required:
      - source
      - target
      - type
      title: GraphEdge
    GraphNode:
      properties:
        id:
          type: string
          title: Id
        label:
          type: string
          title: Label
        properties:
          additionalProperties: true
          type: object
          title: Properties
      type: object
      required:
      - id
      - label
      title: GraphNode
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MitreCoverageItem:
      properties:
        technique_id:
          type: string
          title: Technique Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        tactic:
          anyOf:
          - type: string
          - type: 'null'
          title: Tactic
        alert_count:
          type: integer
          title: Alert Count
      type: object
      required:
      - technique_id
      - name
      - tactic
      - alert_count
      title: MitreCoverageItem
...
    UpsertAlertGraphRequest:
      properties:
        alert_id:
          type: string
          title: Alert Id
        title:
          type: string
          title: Title
        severity:
          type: string
          title: Severity
        mitre_techniques:
          items:
            type: string
          type: array
          title: Mitre Techniques
        host_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Host Id
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
        ioc_values:
          items:
            type: string
          type: array
          title: Ioc Values
      type: object
      required:
      - alert_id
      - title
      - severity
      title: UpsertAlertGraphRequest
    UpsertCaseGraphRequest:
      properties:
        case_id:
          type: string
          title: Case Id
        title:
          type: string
          title: Title
        severity:
          type: string
          title: Severity
        alert_ids:
          items:
            type: string
          type: array
          title: Alert Ids
      type: object
      required:
      - case_id
      - title
      - severity
      title: UpsertCaseGraphRequest
    UpsertHostRequest:
      properties:
        host_id:
          type: string
          title: Host Id
        hostname:
          type: string
          title: Hostname
        ip_address:
          type: string
          title: Ip Address
          default: ''
        os:
          type: string
          title: Os
          default: ''
        criticality:
          type: string
          title: Criticality
          default: medium
      type: object
      required:
      - host_id
      - hostname
      title: UpsertHostRequest
    UpsertUserRequest:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
          default: ''
        department:
          type: string
          title: Department
          default: ''
        risk_score:
          type: number
          title: Risk Score
          default: 0.0
      type: object
      required:
      - user_id
      - username
      title: UpsertUserRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer